Create EventsPageCardList, refactor
